Fix lỗi khẩn cấp ở bài giaiptbac2.py

+++ 23012024/giaiptbac2.py
@@ -1,6 +1,6 @@
-a=int(input("Nhap a:"))
-b=int(input("Nhap b:"))
-c=int(input("Nhap c:"))
+a=float(input("Nhap a:"))
+b=float(input("Nhap b:"))
+c=float(input("Nhap c:"))
 delta=b*b-4*a*c
 if delta<0:
     print("Phuong trinh vo nghiem")
@@ -11,3 +11,4 @@ else:
     x1=(-b+delta**0.5)/(2*a)
     x2=(-b-delta**0.5)/(2*a)
     print("Phuong trinh co hai nghiem x1=",x1,"x2=",x2)
+
